requirements.txt
asgiref==3.5.2
databases==0.6.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
greenlet==1.1.2
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
SQLAlchemy==1.4.37
sqlparse==0.4.2
tzdata==2022.1
